refactor(reader): address review feedback
- Replace dash: [CGFloat] with LinePattern enum for clarity
- Move VerseSelectionStyle to its own file
- Fix "grey" -> "gray" spelling in doc comments
- Revert handleVerseTap fallthrough (onVerseTap behavior
  changes will be addressed in a follow-up PR)
- Remove related onVerseTap selection tests
